Framework tool fixes (#103)
* Small fixes

Apply small fixes as suggested in #100.

* Fix typo in Cros EC driver name argument

* Source profiles to have $PATH set correctly

If `pkexec` runs a program, the enviroment is stripped bare and paths
like `/usr/local/bin` are not in `$PATH`. This attempts at improving the
situation by sourcing the system profile and, if possible, the profile
of the user `pkexec` runs at, i.e. `root`.